auto layout working. logo, despite not appearing on screen during training or game play mode, is blocking the top left button.
the issue was that the width and height sizes were set to constant, (intrinsic values: default vs. placeholder)

and this was not allowing auto resizing to layout elements properly on screen

now all this being said, as i progress in the iOS course, it seems that we will be using swift UI instead of interface builder in most cases, which is more like what i am familiar with.

but at this point, i'm pretty good with auto layout.

so lets push this update, and then solve the logo problem by perhaps updating it's size when not in game over mode